added the rest of chapter 1, 2, and 3 exercises

+++ ch1ex6.py
@@ -2,6 +2,6 @@
 #Eitan
 
 def km_per_hr(distance, time):
-    return distance / float(time / 60)
+    return distance / (time / 60)
 
 print(km_per_hr(10, 45.5))
