+++ ch2ex2.py
@@ -4,6 +4,6 @@
 import math
 
 def sphere_volume(radius):
-    return float(4 / 3) * math.pi * (radius ** 3)
+    return (4 / 3) * math.pi * (radius ** 3)
 
 def price_books_(books):
